no more macros for dlog tests
