--- tests/dlog_solver_tests.rs
@@ -1,97 +1,0 @@
-//! Generic tests for all DLog solver implementations.
-//!
-//! This module tests all DLog algorithms ([BL12], BSGS, BSGS-k) using the
-//! `DlogSolver` trait. For each algorithm, we test bit sizes from 1 to 7,
-//! verifying that the algorithm correctly solves DLog for all values in [0, 2^ℓ).
-
-use curve25519_dalek::constants::RISTRETTO_BASEPOINT_POINT;
-use curve25519_dalek::ristretto::RistrettoPoint;
-use pollard_kangaroo::DlogSolver;
-
-/// Generic test function that tests a DlogSolver implementation for all values
-/// in [0, 2^max_num_bits).
-fn test_all_values<S: DlogSolver>(max_num_bits: u8) {
-    let solver = S::new_and_compute_table(max_num_bits).expect("Failed to create solver");
-
-    let max_value = 1u64 << max_num_bits;
-    let g = RISTRETTO_BASEPOINT_POINT;
-
-    // Start with g^0 = identity, then increment via EC addition: g^(x+1) = g^x + g
-    let mut pk = RistrettoPoint::default(); // identity
-
-    for x in 0..max_value {
-        let result = solver.solve(&pk).expect("Solver returned error");
-
-        assert_eq!(
-            result, x,
-            "Failed for max_num_bits={}, x={}: got {}",
-            max_num_bits, x, result
-        );
-
-        // g^(x+1) = g^x + g
-        pk = pk + g;
-    }
-}
-
-/// Macro to generate tests for a specific DlogSolver implementation.
-/// Tests 1-7 bits exhaustively.
-/// TODO: Don't use a fucking macro for this... Just write a generic function that takes the solver as a template argument and the bit size as a normal argument.
-///   then call this function for all supported solvers in a for loop over the bit sizes. NO MACROS!
-macro_rules! generate_dlog_tests {
-    ($solver:ty, $name:ident) => {
-        mod $name {
-            use super::*;
-
-            #[test]
-            fn test_1_bit() {
-                test_all_values::<$solver>(1);
-            }
-
-            #[test]
-            fn test_2_bit() {
-                test_all_values::<$solver>(2);
-            }
-
-            #[test]
-            fn test_3_bit() {
-                test_all_values::<$solver>(3);
-            }
-
-            #[test]
-            fn test_4_bit() {
-                test_all_values::<$solver>(4);
-            }
-
-            #[test]
-            fn test_5_bit() {
-                test_all_values::<$solver>(5);
-            }
-
-            // #[test]
-            // fn test_6_bit() {
-            //     test_all_values::<$solver>(6);
-            // }
-            //
-            // #[test]
-            // fn test_7_bit() {
-            //     test_all_values::<$solver>(7);
-            // }
-        }
-    };
-}
-
-// Generate tests for BSGS (fastest algorithm)
-#[cfg(feature = "bsgs")]
-generate_dlog_tests!(pollard_kangaroo::bsgs::BabyStepGiantStep, bsgs);
-
-// Generate tests for BSGS-k with K=64
-#[cfg(feature = "bsgs_k")]
-generate_dlog_tests!(pollard_kangaroo::bsgs_k::BabyStepGiantStepK<64>, bsgs_k64);
-
-// Generate tests for BSGS-k with K=256
-#[cfg(feature = "bsgs_k")]
-generate_dlog_tests!(pollard_kangaroo::bsgs_k::BabyStepGiantStepK<256>, bsgs_k256);
-
-// Generate tests for [BL12]
-#[cfg(feature = "bl12")]
-generate_dlog_tests!(pollard_kangaroo::bl12::Bl12, bl12);
